GPT/pyproject.toml
[tool.poetry]
name = "gpt"
version = "0.1.0"
description = ""
authors = ["houston03 <super.avel-2014@yandex.ru>"]


[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
flask-migrate = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
